# swagger/swagger.yaml
...
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Bintracker
  description: Bintracker api
servers:
  - url: "https://staging-api.bintracker.com"
    description: staging
  - url: "https://testing-api.bintracker.com"
    description: testing
paths:
...
  /api/items:
    post:
      summary: POST an item
      tags:
        - items
      description: Post one item.
                   <p />
                   Items are attached to user accounts, so an item posted by
                   one account is not visible or accessible to another account.
                   <p />
                   Items can be added to or removed from bins.
      operationId: ItemController.postItems
      requestBody:
        required: true
        description:
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/item_request'
            example:
              name: Rotini
              amount: 750
              unit_id: 2
      responses:
        201:
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/item_response'

  ###
  # GET /api/items
  #
    get:
      summary: GET page of items
      tags:
        - items
      description: Get one page of user's items.
      operationId: ItemController.getItems
      parameters:
        - name: page
          in: query
          required: false
          description: The page number
          schema:
            type : integer
            format: int64
            minimum: 1
            default: 1
          example: 2
        - name: size
          in: query
          required: false
          description: The number of records per page
          schema:
            type : integer
            format: int64
            minimum: 1
            default: 10
          example: 5
        - name: order
          in: query
          required: false
          description: The column to order on. Use either 'name' or 'created_at'
          schema:
            type : string
            default: name
          example: name 
        - name: sort
          in: query
          required: false
          description: The direction of the sort, either 'asc' or 'desc'
          schema:
            type : string
            default: asc
          example: asc 
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/item_response'
                  links:
                    $ref: '#/components/schemas/pagination'
        404:
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              example:
                error: "Items do not exist"
        403:
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              example:
                error: "Unauthorized"


  ###
  # GET /api/items/units
  #
...
  /api/items/{id}:
    delete:
      summary: DELETE an item
      tags:
        - items
      description: Delete one item. The item must not be in a bin.
      operationId: ItemController.deleteItems
      parameters:
        - name: id
          in: path
          required: true
          description: item id
          schema:
            type : integer
            format: int64
            minimum: 1
          example: 2
      responses:
        201:
          description: CREATED
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    nullable: true
        404:
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              example:
                error: "Item does not exist"
        403:
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              example:
                error: "Cannot delete item that exists in a bin"




  #-----------------------------------------------------
  # BinItems

  ###
  # PUT api/bins/{binId}/items/{itemId}
  #
...
components:
...
  schemas:
...
    item_response:
      type: object
      properties:
        id:
          type: integer
          description: the id of the item
          example: 1
        name:
          type: string
          description: display name of the item
          example: Rotini
        amount:
          type: integer
          description: the quantity of the item. ie. if it's 300g box, the 'amount' would be 300.
          example: 750
        image:
          type: string
          description: the url of the image
          nullable: true
          example: https://example.ca/rotini.jpg
        unit_id:
          type: integer
          description: the id of the units of the amount of this item. ie, if the item is 300 grams, this is 2, the id of the unit 'grams'
          example: 2
        unit:
          type: string
          description: the human-readable unit, ie. 'grams'
          example: grams
        categories:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/item_category_response'
          example:
            - id: 1
              name: food
            - id: 18
              name: Pasta


    ###
    # response: binitem_response
...
    item_category_response:
      type: object
      properties:
        id:
          type: integer
          nullable: true
          description: the id of the category
          example: 2
        name:
          type: string
          nullable: true
          description: the name of the category
          example: Cleaning supplies

     ###
     # response: register_response
...
    item_request:
      type: object
      properties:
        name:
          type: string
        amount:
          type: string
        unit_id:
          type: integer

    ###
    # request: bin_request
...
    pagination:
      type: object
      properties:
        has_more:
          type: boolean
          description: true if there are more pages
          example: true
        current_page:
          type: integer
          description: the number of the page that has been requested
          example: 3
        last_page:
          type: integer
          description: the maximum number of pages in the data set for this size
          example: 8
        current_size:
          type: integer
          description: the number of records per page
          example: 10


    #-----------------------------------------------------
    # error schemas

    ###
    # error: error
    error:
      type: object
      properties:
        error:
          nullable: true
          type: string
        details:
          nullable: true
          type: string